add contributor section
