+++ src/App.js
@@ -1,14 +1,12 @@
 import './App.css';
 import Home from './pages/Home';
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Contributors from './pages/Contributors';
 
 function App() {
   return (
     <BrowserRouter basename={process.env.PUBLIC_URL}>
       <Routes>
         <Route path="/" element={<Home />} />
-        <Route path="/contributors" element={<Contributors />} />
       </Routes>
     </BrowserRouter>
   );
